[PR 10875] [addon/depends] Handle autocrlf for depends
On Windows, if we have addon dependencies, we need to forcefully disable
autocrlf because otherwise the patches will not apply.

Patches are typically in LF format (and the build system automatically adds
--binary if necessary). But the patches will also not apply if the source files
are in CLRF which can happen if the user has core.autocrlf=true in the
global git config.

A better way of fixing would be to pass the config already to clone:
https://gitlab.kitware.com/cmake/cmake/issues/15799